feat: divide to components
